requirements.txt
fastapi
uvicorn
requests
python-dotenv
PyGithub
Pillow
pytesseract
python-multipart